requirements.txt
django
easy-thumbnails
django-ckeditor

#Django Rest Framework
#djangorestframework
#django-tagging
#django-filter
#django-pygments
